Move batch queuers to namespace
